Refactoring - common csv validator code moved to separate file
dcat_test & usmart_test both use the same csv validator test - so it makes sense to group this code in a single common file.

Testing:
ran pytest 100% pass
23 passed in 0.58s
